requirements/dev_ubuntu_cpu.txt
-r dev_base.txt
http://download.pytorch.org/whl/cpu/torch-0.3.1-cp36-cp36m-linux_x86_64.whl 
torchvision
tensorflow==1.6
-e .
